feat: optimize house robber solution with constant space
This commit refactors the house robber solution to use constant space instead of linear space. The previous solution used two variables to store the maximum amount that could be robbed up to the previous two houses. This solution uses two variables to store the maximum amount that could be robbed up to the previous two houses, but it does so in a more efficient way.